requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.0
rsa==4.8
six==1.16.0
Telethon==1.24.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
yarl==1.7.2
